kernels: every flat-indexed kernel computes its offsets from a 64-bit program id — 152 sites in 105 files; a 2.25e9-element tensor was addressed through a wrapped int32 offset (register 58, the class)
With the GEMM fixed (90fefd4) Mochi's op-by-op run went past mm::1 and died
at aten.add::14 — the broadcast add, whose sibling had been widened by hand
in May (Sana 4K VAE) while it had not. The audit found the int32 form
`pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)` at 152 sites in 105 kernels:
fill, the strided copy that materialises a view, the elementwise, reduction,
index, upsample and pixel-shuffle families. All widened at the program id
(`pid.to(tl.int64) * BLOCK_SIZE + ...`, FlagGems #6083's form).

The first form tried — casting after the product — was wrong and its
source-reading gate was green over it: the product had wrapped before the
cast, and the 2.25e9-element fill still faulted (06:31 UTC). The gate
(tests/unit/kernels/test_flat_offsets_are_64_bit.py) now refuses both forms
for any variable name and both BLOCK spellings; the GPU test beside it
(test_a_flat_kernel_beyond_two_billion_elements.py: ones, a transposed
materialisation and an add over 2 252 800 000 elements, the elements before,
at and past 2^31 read back) was RED on 90fefd4 and is GREEN now (2.8 s, card 2).

Measured, card 2, NBX_DISABLE_AUTOTUNE=1, before (90fefd4) vs after x3:
TinyLlama, Kokoro, Sana MultiLing (4 steps), whisper-large-v3-turbo —
byte-identical; warm timings within 3 % (5.56/5.31-5.43, 4.93/4.94-5.19,
31.5/31.4-32.3, 5.19/5.07-5.20 s). Kernel suite on card 3, pinned:
2 failed, 908 passed, 109 skipped, 2 deselected, 3 warnings in 1202.48s (0:20:02) — the two failures (test_prefill_determinism 16g/32g routes) fail
identically on 90fefd4 under the same pin and pass unpinned on this tree:
the pinned-ordinal class (D-TEST-PINNED-ORDINAL), not this change.

+++ src/neurobrix/kernels/ops/adaptive_avg_pool2d.py
@@ -48,7 +48,7 @@ def adaptive_avg_pool2d_kernel(
     corresponding adaptive window in the input.
     """
     pid = tl.program_id(0)
-    idx = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    idx = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = idx < n_elements
 
     # Decompose flat index -> (n, c, oh, ow)
