+++ src/neurobrix/kernels/ops/binary_strided.py
@@ -29,7 +29,7 @@ def add_strided_nd_kernel(
 ):
     """out = x + alpha * y, x and y read by their strides, out contiguous."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
@@ -60,7 +60,7 @@ def mul_strided_nd_kernel(
 ):
     """out = x * y, x and y read by their strides, out contiguous."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
@@ -92,7 +92,7 @@ def sub_strided_nd_kernel(
 ):
     """out = x - alpha * y, x and y read by their strides, out contiguous."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
@@ -123,7 +123,7 @@ def div_strided_nd_kernel(
 ):
     """out = x / y, x and y read by their strides, out contiguous."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
@@ -154,7 +154,7 @@ def add_scalar_strided_nd_kernel(
 ):
     """out = x + scalar, x read by its strides, out contiguous (add_scalar_kernel's expression)."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
@@ -181,7 +181,7 @@ def mul_scalar_strided_nd_kernel(
 ):
     """out = x * scalar, x read by its strides, out contiguous (mul_scalar_kernel's expression)."""
     pid = tl.program_id(0)
-    offsets = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
     x_off = tl.zeros_like(offsets)
