+++ src/neurobrix/kernels/ops/copy_op.py
@@ -27,7 +27,7 @@ def copy_kernel(
     incident). NaN and true ±inf inputs pass through unchanged.
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     data = tl.load(src_ptr + offset, mask=mask)
