+++ src/neurobrix/kernels/ops/dtype_convert.py
@@ -24,7 +24,7 @@ def bf16_to_fp16_kernel(
     The GPU handles fp32→fp16 truncation natively (with rounding).
     """
     pid = tl.program_id(0)
-    offs = pid * BLOCK + tl.arange(0, BLOCK)
+    offs = pid.to(tl.int64) * BLOCK + tl.arange(0, BLOCK)   # 64-bit from the program id (register 58)
     mask = offs < N
 
     # Load bf16 raw bits (stored as int16, same 2-byte layout)
