+++ src/neurobrix/kernels/ops/index_put_op.py
@@ -61,7 +61,7 @@ def index_put_kernel(
     emitted.
     """
     pid = tl.program_id(0)
-    off = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    off = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = off < N
 
     s = off // T                       # which index entry (row in idx)
