+++ src/neurobrix/kernels/ops/lerp.py
@@ -21,7 +21,7 @@ def lerp_tensor_forward_kernel(
 ):
     """Lerp with tensor weight: numerically stable two-branch formula."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     a = tl.load(input_ptr + offset, mask=mask)
@@ -46,7 +46,7 @@ def lerp_scalar_head_kernel(
 ):
     """Lerp with scalar weight < 0.5: a + w*(b-a)."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     a = tl.load(input_ptr + offset, mask=mask)
@@ -66,7 +66,7 @@ def lerp_scalar_tail_kernel(
 ):
     """Lerp with scalar weight >= 0.5: b - (b-a)*(1-w)."""
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     a = tl.load(input_ptr + offset, mask=mask)
