+++ src/neurobrix/kernels/ops/pixel_shuffle.py
@@ -42,7 +42,7 @@ def pixel_shuffle_kernel(
     Each thread handles one output element.
     """
     pid = tl.program_id(0)
-    idx = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    idx = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = idx < n_elements
 
     # Decompose flat output index -> (n, c, oh, ow)
@@ -116,7 +116,7 @@ def pixel_shuffle_broadcast_aware_kernel(
     # (e.g. Sana 4Kpx VAE 1*128*4096*4096 = 2^31 — at INT32_MAX
     # boundary; signed int32 multiplication of strides overflows
     # silently and corrupts offsets, producing garbage output).
-    idx = (pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)).to(tl.int64)
+    idx = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = idx < n_elements
 
     ow = idx % OW
