+++ src/neurobrix/kernels/ops/rand_op.py
@@ -24,7 +24,7 @@ def rand_kernel(
         seed: RNG seed (int)
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     r = tl.rand(seed, offset)
@@ -45,7 +45,7 @@ def randn_kernel(
         seed: RNG seed (int)
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
     r = tl.randn(seed, offset)
